pyproject.toml
[tool.poetry]
name = "app-service"
version = "0.1.0"
description = ""
authors = ["Team 11 <a.chowdhury-2@student.tudelft.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^4.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
